miraclectl: add scan commands
Three new commands that control scans on a given link are added:
 1) start-scan: Starts scanning in background for a given link
 2) stop-scan: Stops scanning in background for a given link
 3) scan: Does interactive scanning for a given link

The interactive scanning basically calls start-scan and stop-scan. In
between, it listens for dbus-signals for new peers and displays them if
they are assigned to the given link. This should be used for basic
p2p-device scanning on a link.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>