miraclectl: suppress useless error on scan-stop
If we stop a scan in an async-manner and the object is no longer around,
don't print an error. Only do that if the scan is stopped explicitly by
the user.

This allows us to correctly kill scans during link-removal in a follow-up
patch. In such cases, the scan-stop may fail of the dbus object has
already been removed.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>