miraclectl: hook up connection API
Add commands for allow/reject/connect/disconnect and parse all events
properly. Notify the user of any important changes so they can react to
it.

Note that Peer.Connect() is currently not implemented and will stay this
way until we have the dhcp-server and wfd-source ready (or at least
planned).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>