miracled: skip run-time device-monitor if running unmanaged
If we get ADD events from udev for a new link but run in unmanaged mode,
skip the event. It is fairly unlikely that another manager has already
started wpa_supplicant on it (and even if, it would be racy).

Hence, lets not support hotplugging in these cases until external managers
provide us a notification API.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>