test: make wpas tests safe to run in parallel
If you run "make check -jX", the test-suites will run in parallel. This
currently works fine, but someone might copy the wpas tests into something
more complex. In that case, the /tmp/XY files will be shared, which breaks
the test-suite. Therefore, use pid-suffixes so these files are exclusive
to the given test-process.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>