Use a guard to prevent dependency on pkg-config when it is not available
I picked the methodology off of https://github.com/curl/curl/issues/972.
The benefit here is that if someone does not have pkg-config (we know this
is unlikely anyways), their configuration script just won't end up with the
checks.

This provides a patch similar to the one requested in issue #120.